Added the token to the code metadata in case we can't evaluate it.
I don't know what would happen if we detect some potential clojure code but then when we try to evaluate it we find out it doesn't make sense as code. I suppose it would depend on the type of error but one way of dealing with it would be to just treat the value as normal text, so that's why I decided to include the token as metadata: to be able to get the actual source behind it.