Changed project.clj to only AOT the ssgr.core namespace.
This fixes a problem with the ssgr.parser that was breaking the uberjar compilation.
For some reason, when compiling AOT all the namespaces it didn't find the as-parser implementation for ssgr.parser.ClojureParser, which doesn't make sense since ClojureParser is a subclass of petitparser.parsers.Parser so it should have the as-parser already implemented.
Anyway, I don't quite understand how all that works but I found out I don't really need to AOT compile all the namespaces, just the entry point is enough and it works pretty much the same.
If there is some problem with this I guess I'll find out in the future. We'll see...