The parser can now evaluate clojure code.
I would have preferred to keep parsing and evaluation completely separate. However, I had to do it this way because in order to know if the clojure code is valid or not I need to evaluate it. And if the code throws an error then we can still parse it as markdown.