Changed the way the parser works, which fixes some problems.
I basically made the parser work (for the most part) with a single character stream which gets passed around through different parsers. This fixes an issue I had before in which clojure code including blank lines would not be parsed correctly because the paragraph parsing was taking precedence.
I think the code is a little simpler now, but in any case I'll have to clean it up when I finished implementing the markdown syntax. There are a lot of things missing still but at least we now support everything that we did before the refactor :)